backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
PyJWT==2.9.0
motor==3.3.2
pymongo==4.5.0
httpx==0.27.2
orjson==3.10.7
email-validator==2.1.1
python-multipart==0.0.9
openai==1.40.6
pytest==8.3.2
redis==5.0.1
websockets==12.0
ollama==0.3.3
aiohttp==3.9.1
wyoming==1.5.3
stripe==13.2.0
psutil==5.9.8
